# docs/swagger.yaml
definitions:
  http.ErrorResponse:
    properties:
      error: {}
      status:
        type: string
    type: object
...
  teklif_internal_api_models.OfferResp:
    properties:
      commission_agency:
        type: number
      commission_agent:
        type: number
      description:
        type: string
      discount:
        type: number
      expiry_date:
        type: string
      id:
        type: string
      media:
        items:
          type: string
        type: array
      price_cash:
        type: number
      price_installment:
        type: number
      status:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /offers/active:
    get:
      consumes:
      - application/json
      description: Retrieve a list of active offers that are approved and not expired
      produces:
      - application/json
      responses:
        "200":
          description: List of active offers
          schema:
            items:
              $ref: '#/definitions/teklif_internal_api_models.OfferResp'
            type: array
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: No offers found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Get active offers
      tags:
      - offers
...
  /offers/week-expired:
    get:
      consumes:
      - application/json
      description: Retrieve a list of offers that are marked as "Expired" and expired
        in the last 7 days.
      produces:
      - application/json
      responses:
        "200":
          description: List of expired offers
          schema:
            items:
              $ref: '#/definitions/teklif_internal_api_models.OfferResp'
            type: array
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: No offers found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Get list of offers with "Expired" status that have expired in the last
        7 days.
      tags:
      - offers
swagger: "2.0"